==> 01_Load_and_inspect.ipynb <==
# Data preparation and transformation exercise

### Part I - Load and Inspect

The objective of this exercise is to practice various steps of data preprocessing and feature engineering.

The scenario is the preparation of data for a ML multilinear regressions.

The dataset used is the "Climate Weather Surface of Brazil - Hourly", wich is available at <a href="https://www.kaggle.com/PROPPG-PPG/hourly-weather-surface-brazil-southeast-region?select=make_dataset.py">Kaggle</a>.

It contains hourly climate data taken from 122 weather stations in Brasil between 2000 and 2021.

**Steps:**
1. <a href="#Load-data">Load data</a>
2. <a href="#Inspect-data">Inspect data</a>
3. Format features
4. Clean messy data
5. Remove duplicate values
6. Treat missing values
7. Imputation
8. Remove strongly correlated features
9. Remove outliers
10. Aggregate features
11. Encode categorical features
12. Feature scaling
13. Dimensionality reduction and feature decomposition
14. Sample and balance

## About the dataset

| Original position | Column Name | Description | Position on load |
|--|--|--|--|
| 0 | date | date (YYYY-MM-DD) | not loaded |
| 1 | hour | hour (HH:00) | not loaded |
| NA | full_time | date + time | 0 |
| 2 | total precipitation (mm) | Amount of precipitation in millimetres (last hour) | 1 |
| 3 | atmospheric pressure at station height (mb) | Atmospheric pressure at station level (mb) | 2 |
| 4 | atmospheric pressure max. in the previous hour (mb) | Maximum air pressure for the last hour in hPa to tenths | 3 |
| 5 | atmospheric pressure min. in the previous hour (mb) | Minimum air pressure for the last hour in hPa to tenths | 4 |
| 6 | radiation (kj/m2) | Solar radiation KJ/m2 | 5 |
| 7 | air temperature - dry bulb (¬∞c) | Air temperature (instant) in celsius degrees | 6 |
| 8 | dew point temperature (¬∞c) | Dew point temperature (instant) in celsius degrees | 7 |
| 9 | max. temperature in the previous hour (¬∞c) | Maximum temperature for the last hour in celsius degrees | 8 |
| 10 | min. temperature in the previous hour (¬∞c) | Minimum temperature for the last hour in celsius degrees | 9 |
| 11 | dew temperature max. in the previous hour (¬∞c) | Maximum dew point temperature for the last hour in celsius degrees | 10 |
| 12 | dew temperature min. in the previous hour (¬∞c) | Minimum dew point temperature for the last hour in celsius degrees | 11 |
| 13 | relative humidity max. in the previous hour (%) | Maximum relative humid temperature for the last hour in % | 12 |
| 14 | relative humidity min. in the previous hour (%) | Minimum relative humid temperature for the last hour in % | 13 |
| 15 | air relative humidity (%) | Relative humid in % (instant) | 14 |
| 16 | wind direction (¬∞ (gr)) | Wind direction in radius degrees (0-360) | 15 |
| 17 | wind rajada maxima (m/s) | Wind gust in metres per second | 16 |
| 18 | wind speed (m/s) | Wind speed in metres per second | 17 |
| 19 | region | Brazilian geopolitical regions | 18 |
| 20 | state | State (Province) | 19 |
| 21 | station | Name station (usually city location or nickname) | 20 |
| 22 | station_code | Station number (INMET number) for the location | 21 |
| 23 | latitude | Latitude | 22 |
| 24 | longitude | Longitude | 23 |

## Load data

import pandas as pd
import numpy as np
import pickle

# automatic feature type detection is resource intensive, so we set the types manually.
dataset = pd.read_csv('./dataset/archive/central_west.csv',
                      index_col="index",
                      parse_dates={'full_time' : [1, 2]},
                      dtype={'Data': object,
                             'Hora': object,
                             'PRECIPITAÇÃO TOTAL, HORÁRIO (mm)': np.float64,
                             'PRESSAO ATMOSFERICA AO NIVEL DA ESTACAO, HORARIA (mB)': np.float64,
                             'PRESSÃO ATMOSFERICA MAX.NA HORA ANT. (AUT) (mB)': np.float64,
                             'PRESSÃO ATMOSFERICA MIN. NA HORA ANT. (AUT) (mB)': np.float64,
                             'RADIACAO GLOBAL (Kj/m²)': np.int64,
                             'TEMPERATURA DO AR - BULBO SECO, HORARIA (°C)': np.float64,
                             'TEMPERATURA DO PONTO DE ORVALHO (°C)': np.float64,
                             'TEMPERATURA MÁXIMA NA HORA ANT. (AUT) (°C)': np.float64,
                             'TEMPERATURA MÍNIMA NA HORA ANT. (AUT) (°C)': np.float64,
                             'TEMPERATURA ORVALHO MAX. NA HORA ANT. (AUT) (°C)': np.float64,
                             'TEMPERATURA ORVALHO MIN. NA HORA ANT. (AUT) (°C)': np.float64,
                             'UMIDADE REL. MAX. NA HORA ANT. (AUT) (%)': np.int64,
                             'UMIDADE REL. MIN. NA HORA ANT. (AUT) (%)': np.int64,
                             'UMIDADE RELATIVA DO AR, HORARIA (%)': np.int64,
                             'VENTO, DIREÇÃO HORARIA (gr) (° (gr))': np.int64,
                             'VENTO, RAJADA MAXIMA (m/s)': np.float64,
                             'VENTO, VELOCIDADE HORARIA (m/s)': np.float64,
                             'region': object,
                             'state': object,
                             'station': object,
                             'station_code': object,
                             'latitude': np.float64,
                             'longitude': np.float64,
                             'height': np.float64},
                      names=["index",
                             "date", 
                             "time", 
                             "precipitation", 
                             "pressure", 
                             "pressure_max", 
                             "pressure_min", 
                             "solar_radiation", 
                             "air_temperature", 
                             "dp_temperature", 
                             "air_temp_max", 
                             "air_temp_min", 
                             "dp_temp_max", 
                             "dp_temp_min", 
                             "rel_hum_max", 
                             "rel_hum_min", 
                             "Rel_humidity", 
                             "wind_direction", 
                             "wind_gust", 
                             "wind_speed",
                             "region", 
                             "state", 
                             "station_name", 
                             "station_id", 
                             "Latitude", 
                             "Longitude", 
                             "Elevation"],
                     header = 0)

pickle.dump(dataset, open("raw_dataset.pkl", "wb"))

## Inspect data

dataset

### Determine/confirm initial feature types

dataset.info()

The types are as requested, so no strings existed in float or integer features.

dataset.nunique(axis=0)

The features 'region', 'state', 'station_name' / 'station_id' are categorical.

We have data from 116 different weather stations, but one is missing its ID.

print ("We have data from {} , up to {}.".format(str(dataset.full_time.min()), str(dataset.full_time.max())))

first_day = dataset[dataset.full_time == '2000-05-07 00:00:00'].full_time.value_counts()
last_day = dataset[dataset.full_time == '2021-04-30 23:00:00'].full_time.value_counts()
print ("We have {} data points on the first day of the dataset and {}  in the last.".format(str(first_day[0]),
                                                                                            str(last_day[0])))

Let's see the number of stations per year.

# we'll change the type of 'date' feature later. For now we'll just extract the year
dataset['year'] = dataset.full_time.dt.year

dataset.groupby('year')['station_name'].nunique()

This dataset started with a signle station and had 113 stations by 2021.

dataset.index.is_unique

There are duplicates in the index data loaded from the CSV file.

dataset[dataset.index==0].head()

# Let's reset the index
dataset.reset_index(inplace=True)
dataset.drop(columns = ["index"], inplace=True)

dataset

pickle.dump(dataset, open("newindex_dataset.pkl", "wb"))

### Inspect categorical features

# Determine which Brazilian states have values
states = dataset.state.unique()
print(states)

# Determine number of weather stations per state
dataset.groupby('state')['station_name'].nunique()

Let's check if there are different values for latitude, longitude and elevation for the same station.

dataset.groupby("station_name")['Latitude'].nunique().max()

dataset.groupby("station_name")['Longitude'].nunique().max()

dataset.groupby("station_name")['Elevation'].nunique().max()

Yes, we have different values for some stations.

# let's look at an example
duplicated_station_name = dataset.groupby("station_name")['Latitude'].nunique().index[0]
duplicated_station_name

dataset.loc[dataset.station_name == duplicated_station_name, ['Latitude', 
                                                              'Longitude', 
                                                              'Elevation']].drop_duplicates()

These two values for Elevation must be due to a correction of the elevation value associated with the station (assuming the station wasn't raised or lowered by eight meters.

### Find missing values

dataset.isnull().sum()

There appear to be no missing values, but we saw earlier that we had one fewer station_id than station_name.

# Let's see if the same station has multiple IDs
station_list = dataset.iloc[:,[20,21]].value_counts().to_frame()
station_list.reset_index(inplace=True)
station_list.drop(columns=0, inplace=True)

station_list

station_list.sort_values(by='station_name')

station_list.station_id.value_counts()

The station with 'station_id' == A927 has two different names ("BRASNORTE (NOVO MUNDO)" and "BRASNORTE (MUNDO NOVO)"). We'll standardize that later.
