==> 02_Format_Clean_Remove_duplicates.ipynb <==
# Data preparation and transformation exercise

## Part II - Format, Clean and Remove duplicates

The objective of this exercise is to practice various steps of data preprocessing and feature engineering.

The scenario is the preparation of data for a ML multilinear regressions.

The dataset used is the "Climate Weather Surface of Brazil - Hourly", wich is available at <a href="https://www.kaggle.com/PROPPG-PPG/hourly-weather-surface-brazil-southeast-region?select=make_dataset.py">Kaggle</a>.

It contains hourly climate data taken from 122 weather stations in Brasil between 2000 and 2021.

**Steps:**
1. Load data
2. Inspect data
3. <a href="#Format-features">Format features</a>
4. <a href="#Clean-messy-data">Clean messy data</a>
5. <a href="#Remove-duplicate-values">Remove duplicate values</a>
6. Treat missing values
7. Imputation
8. Remove strongly correlated features
9.  Remove outliers
10. Aggregate features
11. Encode categorical features
12. Feature scaling
13. Dimensionality reduction and feature decomposition
14. Sample and balance

import pandas as pd
import numpy as np
import pickle
dataset = pickle.load(open("newindex_dataset.pkl", "rb"))

## Format features

Let's look at the first row, to get a feeling for how the features are presented.

dataset.iloc[0]

Let's check maximum and minimum values for each feature, to tee if they are within reasonable bounds

dataset.max()

The maximum values seem ok for these feature types

dataset.min()

Clearly we have -9999 replacing missing values.

## Clean messy data

As we saw before, the values for Latitude, Longitude and Elevation in each station are not unique, but as they are categorical features that correspond to a same station, like 'region' and 'state', so we'll discard them, as well as the station_id, and use the categorical feature station_name (one hot encoded).

This still leaves us with the two different values for station_name of the same station, so before discarding station_id, we'll unify the name for station with 'station_id' == A927 as "BRASNORTE (NOVO MUNDO)"

dataset.loc[dataset.station_id == 'A927', 'station_name'] = "BRASNORTE (NOVO MUNDO)"

dataset.station_name.loc[dataset.station_id == 'A927'].unique()

dataset.drop(columns = ['station_id', 
                        'Latitude', 
                        'Longitude', 
                        'Elevation', 
                        'year', 
                        'region', 
                        'state'], inplace=True)

pickle.dump(dataset, open("clean_dataset.pkl", "wb"))

## Remove duplicate values

dataset.duplicated().value_counts()

There are no duplicates in the dataset.
