==> 03_Missing_Values_and_Imputation.ipynb <==
# Data preparation and transformation exercise

## Part III - Missing values and Imputation

The objective of this exercise is to practice various steps of data preprocessing and feature engineering.

The scenario is the preparation of data for a ML multilinear regressions.

The dataset used is the "Climate Weather Surface of Brazil - Hourly", wich is available at <a href="https://www.kaggle.com/PROPPG-PPG/hourly-weather-surface-brazil-southeast-region?select=make_dataset.py">Kaggle</a>.

It contains hourly climate data taken from 122 weather stations in Brasil between 2000 and 2021.

**Steps:**
1. Load data
2. Inspect data
3. Format features
4. Clean messy data
5. Remove duplicate values
6. <a href="#Treat-missing-values">Treat missing values</a>
7. <a href="#Imputation">Imputation</a>
8. Remove strongly correlated features
9. Remove outliers
10. Aggregate features
11. Encode categorical features
12. Feature scaling
13. Dimensionality reduction and feature decomposition
14. Sample and balance

import pandas as pd
import numpy as np
import pickle
dataset = pickle.load(open("clean_dataset.pkl", "rb"))

## Treat missing values

As seen above, missing values were replaced with -9999 (solar_radiation) or -9999.0 (other features).

If all sensor values are simultaneously missing, we will discard the row.

duplicates = dataset.loc[
    (dataset.precipitation == -9999.0) & 
    (dataset.pressure == -9999.0) & 
    (dataset.pressure_max == -9999.0) & 
    (dataset.pressure_min == -9999.0) & 
    (dataset.solar_radiation == -9999) & 
    (dataset.air_temperature == -9999.0) & 
    (dataset.dp_temperature == -9999.0) & 
    (dataset.air_temp_max == -9999.0) & 
    (dataset.air_temp_min == -9999.0) & 
    (dataset.wind_gust == -9999.0) & 
    (dataset.wind_speed == -9999.0)].index
print("The dataset has {} empty measurements".format(len(duplicates)))

# Let's drop the duplicates
dataset.drop(index=duplicates, inplace=True)

The empty readings are gone. Next we will replace individual missing readings (-9999) with np.nan, to facilitate their tratment.

def replace_nines(col_name :str):
    """
    Replace the values -9999 in the column specified of the (global) dataset with np.nan
    """
    global dataset
    
    dataset.loc[dataset[col_name] == -9999.0, col_name] = np.nan

numeric_features = ["precipitation",
                    "pressure",
                    "pressure_max",
                    "pressure_min",
                    "solar_radiation",
                    "air_temperature",
                    "dp_temperature",
                    "air_temp_max",
                    "air_temp_min",
                    "dp_temp_max",
                    "dp_temp_min",
                    "rel_hum_max",
                    "rel_hum_min",
                    "Rel_humidity",
                    "wind_direction",
                    "wind_gust",
                    "wind_speed"]

for i in numeric_features:
    replace_nines(i)

pickle.dump(dataset, open("with_nan_dataset.pkl", "wb"))

## Imputation

Next we will use the mean value of the feature in each station to replace the missing values of that feature.

**Warning:** we should not calculate the mean of the training dataset using values from the test dataset. However, in order to keep this exercise simple, we will ignore this and impute the missing values before spliting the dataset.

# count missing values per feature
print(dataset.isnull().sum())

# we'll generate a dataset (frame) with the data of each station
# and use them to calculate the mean to imputate missing values
frames = []

# number of numeric features present in the dataset
num_features = 17

# position offset for the first numeric feature within datates
feature_offset = 1

# list with names of all the stations
stations = dataset['station_name'].unique().tolist()

a = 1
for i in stations: # iterate stations
    # create dataset with data of that station only
    df_station = dataset[dataset['station_name']== i].copy()
    print('processing station {} {}'.format(str(a),i))
    a +=1
    for j in range(feature_offset, feature_offset + num_features): # iterate features
        df_station[dataset.columns[j]].fillna(df_station[dataset.columns[j]].mean(),inplace = True)
    frames.append(df_station)

len(frames)

# recreate the dataset from the cleaned frames
dataset = pd.concat(frames)

pickle.dump(dataset, open("imputed_dataset.pkl", "wb"))

# confirm that we don not have missing values
print(dataset.isnull().sum())
